Find Pivot index
Here pivot index is the index where the sum of all the numbers to the left of the index is equal to the sum of all the numbers to the right of the index.
If no such index exists, we should return -1. If there are multiple pivot indexes, we should return the left-most pivot index.